add projects deletion
