+++ src/components/project/projectEventListeners.js
@@ -1,7 +1,8 @@
-import helpers from '../../helpers';
 import elements from '../../elements';
 import projectHelpers from './projectHelpers';
 import showItemsFirst from '../items/showItemsFirst';
+import displayProjectForm from './displayProjectForm';
+import addNewProject from './addNewProject';
 
 const addItemGeneratorToProject = () => {
   const { projectUl } = elements.ele();
@@ -10,8 +11,8 @@ const addItemGeneratorToProject = () => {
 
 const addEventListeners = () => {
   const { projectContainer, showFormBtn, projectUl } = elements.ele();
-  projectContainer.addEventListener('submit', projectHelpers.createNewProject);
-  showFormBtn.addEventListener('click', helpers.displayForm);
+  projectContainer.addEventListener('submit', addNewProject.addNewProject);
+  showFormBtn.addEventListener('click', displayProjectForm.displayForm);
   const btns = projectUl.querySelectorAll('button');
   btns.forEach((btn) => {
     btn.addEventListener('click', projectHelpers.deleteProject);
